refactor: move tokenization request-id types to the integration module
