+++ src/unwrapped_equity_recovery/job.rs
@@ -51,7 +51,8 @@ use crate::equity_redemption::{EquityRedemption, RedemptionAggregateId};
 use crate::inventory::BroadcastingInventory;
 use crate::inventory::view::{InFlightEquityLocation, InventoryView};
 use crate::rebalancing::trigger::{GuardState, claim_guard_for_recovery_or_orphan};
-use crate::tokenized_equity_mint::{IssuerRequestId, TokenizedEquityMint};
+use crate::tokenization::IssuerRequestId;
+use crate::tokenized_equity_mint::TokenizedEquityMint;
 
 /// Apalis queue type for [`UnwrappedEquityRecoveryJob`].
 pub(crate) type UnwrappedEquityRecoveryJobQueue = JobQueue<UnwrappedEquityRecoveryJob>;
@@ -709,10 +710,9 @@ mod tests {
     use crate::rebalancing::equity::{CrossVenueEquityTransfer, EquityTransferServices};
     use crate::rebalancing::trigger::InProgressGuard;
     use crate::tokenization::Tokenizer;
+    use crate::tokenization::issuer_request_id;
     use crate::tokenization::mock::{MockCompletionOutcome, MockDetectionOutcome, MockTokenizer};
-    use crate::tokenized_equity_mint::{
-        TokenizedEquityMint, TokenizedEquityMintCommand, issuer_request_id,
-    };
+    use crate::tokenized_equity_mint::{TokenizedEquityMint, TokenizedEquityMintCommand};
     use crate::vault_lookup::MockVaultLookup;
 
     use super::super::aggregate::UnwrappedEquityRecoveryServices;
